Tactical Asset System
The TAA system is based on an underlying model Bayesian Model, BayesianK2.pkl. This is included in the GitHub Reporsitory. 
Operating Instructions
Download the Bayesian Model 'BayesianK2.pkl' and store in a path on your computer 'C:/MScFE/Capstone/Models/BayesianK2.pkl'
Download Notebook 'TAA_Bayesian'. Run
Enter the Date in 'YYYY-MM-DD' format, on which we want to generate a Signal to perform Tactical Asset Allocation. Preferred as 1st of a Month
# Condition 1. If the current date is passed (not more than 3 days), then use 1st of that month as End Date. Current Date = 3rd Sept, then End Date is 1st Sept
# Act: In this condition, if the Forecast Signal is 1 Rebalance to TAA, else stay with SAA.
# Condition 2. If current data is less than 1st of a month when the Signal should be generated. We can use the Signal date as End Date
# Run this system again close to the Date of Signal and Act as per above